requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.29
numpy==1.22.3
pandas==1.4.2
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
schedule==1.1.0
six==1.16.0
urllib3==1.26.9
